requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
pydantic>=2.0.0
openai>=1.10.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
